+++ aws/container-linux/kubernetes/workers/workers.tf
@@ -41,7 +41,7 @@ resource "aws_autoscaling_group" "workers" {
 
 # Worker template
 resource "aws_launch_configuration" "worker" {
-  image_id      = "${data.aws_ami.coreos.image_id}"
+  image_id      = "${local.ami_id}"
   instance_type = "${var.instance_type}"
   spot_price    = "${var.spot_price}"